Give image editing and deletion permission to the owner of the image only

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,26 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
+        //Get image entity model from the db
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+
+        //Check if logged in user is the owner of the image.
+        //'isAllowed' will be true if owner of image and logged in user are same.
+        // Otherwise image editing is not allowed
+        Boolean isAllowed = isCurrentUserOwner(image, session);
+
+        if (isAllowed) {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+        else {
+            model.addAttribute("editError", true);
+            return "/images/image";
+        }
+
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,11 +153,24 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
-    }
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+        //Get image entity model from the db
+        Image image = imageService.getImage(imageId);
+        //Check if logged in user is the owner of the image.
+        //'isAllowed' will be true if owner of image and logged in user are same.
+        // Otherwise image deletion is not allowed
+        Boolean isAllowed = isCurrentUserOwner(image, session);
 
+        if (isAllowed) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else {
+            model.addAttribute("image", image);
+            model.addAttribute("deleteError", true);
+            return "/images/image";
+        }
+    }
 
     //This method converts the image to Base64 format
     private String convertUploadedFileToBase64(MultipartFile file) throws IOException {
@@ -187,4 +213,19 @@ public class ImageController {
 
         return tagString.toString();
     }
+
+    private Boolean isCurrentUserOwner(Image image, HttpSession httpSession) {
+        //Get user from the session object
+        User loggedUser = (User) httpSession.getAttribute("loggeduser");
+        //Gent owner of the image
+        User imageOwner = image.getUser();
+        //Check if logged in user's id is equal to the provided image user
+        if(loggedUser.getId() == imageOwner.getId()) {
+            return true;
+        }
+        else {
+            return false;
+        }
+    }
+
 }
